// utils/types.ts
// types.ts
export interface ImageData {
    id: string;
    modelImage: string;
    modelName: string;
    photographerName: string;
    photographerId: string;
}

export interface ImageProps {
    id: number;
    height: string;
    width: string;
    public_id: string;
    format: string;
    blurDataUrl?: string;
}